+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.zora",
+    package = "zorify.mindsgn.studio",
 )
 
 android_resource(
     name = "res",
-    package = "com.zora",
+    package = "zorify.mindsgn.studio",
     res = "src/main/res",
 )
