+++ android/app/src/main/java/zorify/mindsgn/studio/MainActivity.java
@@ -1,4 +1,4 @@
-package com.zora;
+package zorify.mindsgn.studio;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -12,7 +12,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "zora";
+    return "zorify";
   }
 
   /**
