+++ android/app/src/main/java/zorify/mindsgn/studio/newarchitecture/components/MainComponentsRegistry.java
@@ -1,4 +1,4 @@
-package com.zora.newarchitecture.components;
+package zorify.mindsgn.studio.newarchitecture.components;
 
 import com.facebook.jni.HybridData;
 import com.facebook.proguard.annotations.DoNotStrip;
