+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'zora'
+rootProject.name = 'zorify'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
 include ':app'
 includeBuild('../node_modules/react-native-gradle-plugin')
