+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "zora",
-  "displayName": "zora"
+  "name": "zorify",
+  "displayName": "zorify"
 }
