+++ ios/zorify.xcodeproj/xcshareddata/xcschemes/zorify.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "zora.app"
-               BlueprintName = "zora"
-               ReferencedContainer = "container:zora.xcodeproj">
+               BuildableName = "zorify.app"
+               BlueprintName = "zorify"
+               ReferencedContainer = "container:zorify.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "zoraTests.xctest"
-               BlueprintName = "zoraTests"
-               ReferencedContainer = "container:zora.xcodeproj">
+               BuildableName = "zorifyTests.xctest"
+               BlueprintName = "zorifyTests"
+               ReferencedContainer = "container:zorify.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "zora.app"
-            BlueprintName = "zora"
-            ReferencedContainer = "container:zora.xcodeproj">
+            BuildableName = "zorify.app"
+            BlueprintName = "zorify"
+            ReferencedContainer = "container:zorify.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "zora.app"
-            BlueprintName = "zora"
-            ReferencedContainer = "container:zora.xcodeproj">
+            BuildableName = "zorify.app"
+            BlueprintName = "zorify"
+            ReferencedContainer = "container:zorify.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
