+++ ios/zorify/Images.xcassets/AppIcon.appiconset/Contents.json
