+++ ios/zorify/Images.xcassets/Contents.json
