+++ ios/zorifyTests/Info.plist
