+++ ios/zorifyTests/zorifyTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface zoraTests : XCTestCase
+@interface zorifyTests : XCTestCase
 
 @end
 
-@implementation zoraTests
+@implementation zorifyTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
