+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "zora",
+  "name": "zorify",
   "version": "0.0.1",
   "private": true,
   "scripts": {
